Bug "fix".
There is something wrong when calling jfield_eta, but who cares. The
core GIMIC has now been separated into a library, but some interface
work remains. All the rest of GIMIC should be reimplemented in Python,
since 90% of all modifications is in this part of the code (and it's a
bit untidy).